+++ tests/testthat/test-get_ont_lvls.R
@@ -6,10 +6,10 @@ test_that("get_ont_lvls works", {
 
   terms1 <- c(parents,childs)
   #### Using only immediate children ####
-  lvls1 <- get_ont_lvls(terms = terms1)
+  lvls1 <- get_ont_lvls(terms = terms1, absolute = TRUE)
   testthat::expect_length(lvls1, length(terms1))
-  testthat::expect_equal(min(lvls1),0)
-  testthat::expect_equal(max(lvls1),1)
+  testthat::expect_equal(min(lvls1),1)
+  testthat::expect_equal(max(lvls1),14)
 
   #### Using all descendants ####
   terms2 <- ontologyIndex::get_descendants(ontology = hpo,
@@ -18,7 +18,13 @@ test_that("get_ont_lvls works", {
   lvls2 <- get_ont_lvls(terms = terms2)
   testthat::expect_length(lvls2, length(terms2))
   testthat::expect_equal(min(lvls2),0)
-  testthat::expect_equal(max(lvls2),3)
+  testthat::expect_equal(max(lvls2),5)
+
+  #### Using relative levels ####
+  lvls3 <- get_ont_lvls(terms = terms1, absolute = FALSE)
+  testthat::expect_length(lvls3, length(terms1))
+  testthat::expect_equal(min(lvls3),0)
+  testthat::expect_equal(max(lvls3),1)
 
   #### Visual confirmation of correct hierarchy ####
   # ontologyPlot::onto_plot(ontology = hpo,